requirements.txt
streamlit
pandas
numpy
scikit-learn
fastapi
uvicorn
xgboost
imbalanced-learn
psycopg2-binary
python-dotenv
